requirements.txt
alembic==1.7.5
autopep8==1.6.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
protobuf==3.19.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
SQLAlchemy==1.4.29
toml==0.10.2
Werkzeug==2.0.2
WTForms==3.0.1
